+++ site/themes/simple/gulpfile.js
@@ -1,0 +1,23 @@
+//
+// Flextype Gulp.js
+// (c) Sergey Romanenko <https://github.com/Awilum>
+//
+
+var Promise = require("es6-promise").Promise,
+    gulp = require('gulp'),
+    csso = require('gulp-csso'),
+    autoprefixer = require('gulp-autoprefixer'),
+    sass = require('gulp-sass');
+
+gulp.task('process-css', function() {
+    return gulp.src('assets/scss/simple.scss')
+        .pipe(sass().on('error', sass.logError))
+        .pipe(autoprefixer({
+            browsers: ['last 2 versions'],
+            cascade: false
+        }))
+        .pipe(csso())
+        .pipe(gulp.dest('assets/dist/css/'));
+});
+
+gulp.task('default', ['process-css'], function() { });