+++ site/themes/simple/views/partials/footer.php
@@ -1,6 +1,6 @@
 <?php namespace Flextype; ?>
 
-    <?php Themes::template('partials/tail')->display(); ?>
+    <?php Themes::view('partials/tail')->display(); ?>
 
     <div class="powered">
       Powered by <a href="http://flextype.org">Flextype</a>